seongeun, fix: PriceInput 컴포넌트로 분리 #117
